meson: allow use of installed uapi headers
Commit c70a9d9d536aefcf6574328aaac0fcc86f9ab0f7 added a test for
$UAPI/include/uapi/evl that was meant to ensure the given $UAPI path was
correct.  That test meant that $UAPI had to point to a linux-evl source
directory and could not point to installed headers in e.g. /usr/include.
The script had previously allowed the use of installed headers, using
the existence of $UAPI/Kbuild as an heuristic to decide whether $UAPI
pointed to linux-evl source or to installed headers.

Rewrite the script to check for the existence of all three required
directories in either location.  This allows the use of installed
headers again.

Fixes: c70a9d9 (build: tighten consistency check for uapi)
Signed-off-by: Evan Gates <evan.gates@gmail.com>