tests: monitor-event: check for WOLI management
Until [1] went to the EVL kernel tree, we would receive the LKSLEEP
notification when attempting to sleep on the tested event guarded by a
mutex, unless T_WOLI is forcibly disabled, which we used to do. This
whole logic was pretty silly, since we should not have been subject to
receive this error notification in the first place, which [1] fixes.

With this fix in place, we can take the opposite steps: force enable
WOLI+HMSIG, to make sure we never get this notification while
attempting to sleep on the event.

The side-effect is that older kernels which do not have that fix will
happily notify the new test code, leading to the related assertion to
fail. For good measure, give a clear hint about the reason why the
test may now fail with older kernels.

[1] evl/wait: prevent LKSLEEP HM notification upon sleep on a gated event

Signed-off-by: Philippe Gerum <rpm@xenomai.org>