tests: sched-quota-accuracy: print out the percentage of fulfillment
The quota scheduling accuracy may yield varying results, which clearly
depend on the execution environment; as a consequence, it is difficult
for the test code to figure out whether it did succeed or failed.

Instead, print out the percentage of the quota which has been actually
received for the user to interpret, determining whether the outcome
was ok or not, considering the test conditions.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>