utils: test: account for status-less tests
Some tests might not be able to return a definitive ok/broken status,
but rather some information the user would interpret for determining
whether the outcome was ok or not, considering the test conditions.

Typically, measuring the quota scheduling accuracy may yield varying
results, which clearly depend on the execution environment.

The standard output of a test is now always captured, printed to the
terminal if the status code was EXIT_NO_STATUS.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>