treewide: sanitize uapi/ handling
Referring to a local custom copy of the kernel headers via the -Duapi
option makes sense in maintainer mode (only), when rebuilding a
compilation toolchain to include an updated set of exported kernel
headers would be overkill.

In the wake of the uapi/ sanitization in the EVL kernel tree, fix up
the include directives according to those changes.

Basically, stop mentioning the uapi/ root in header paths, so that we
can refer to them the same way whether they belong to a pre-installed
native or cross-compilation toolchain, or a local development kernel
tree.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>