build: drop uapi installation script
Installing the kernel uapi/ files as part of the libevl artefacts is a
relic from the dark ages, when the target toolchain would not have a
copy of those. In this day and age, we build proper toolchains which
do include a copy of the exported uapi/ headers.

For the maintainer mode where using a local copy of the kernel uapi/
headers in place is handier, we have the -Duapi switch.

Drop copy-uapi.sh and the related meson rules.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>