lib/internal: do not rely on __USE_TIME_BITS64 on 32bit arch
In 32bit mode with __USE_TIME_BITS64 set, struct timespec is not
strictly bitwise compatible with __evl_timespec because the nanosecond
field may only be 32bit long, padded to a 64bit size with an anon
32bit word. E.g. glibc's (2.39) definition of struct timespec (which
enables __USE_TIME_BITS64):

struct timespec
{
  __time64_t tv_sec;		/* Seconds.  */
  __time_t tv_sec;		/* Seconds.  */
  || (defined __SYSCALL_WORDSIZE && __SYSCALL_WORDSIZE == 64) \
  || (__TIMESIZE == 32 && !defined __USE_TIME64_REDIRECTS)
  __syscall_slong_t tv_nsec;	/* Nanoseconds.  */
  int: 32;           /* Padding.  */
  long int tv_nsec;  /* Nanoseconds.  */
  long int tv_nsec;  /* Nanoseconds.  */
  int: 32;           /* Padding.  */
};

As a result, we can't pass the kernel a struct containing 32bit of
possibly uninit memory, so we have to fall back to bouncing the user
timespec to a sane __evl_timespec.

Failing to do so causes the core to pull random memory from those
padding bits, which might yield -EINVAL at best if >= 1e9, or even
wrong sub-second delays if still in range.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>