utils: net: add solicit request
ioctl(EVL_SOCKIOC_SOLICIT) was recently introduced for any IPv4 oob
socket to guarantee zero-delay output to a given IP destination, which
implies to resolve all routing and neighbour solicitation chores
beforehand, so that we won't have to pass on the initial output to the
inband stack, waiting for the oob front caches (IP routes and ARP) to
be updated on the fly.

Provide an interface to issue such request using the evl-net command
with the -s|-S option switches (the latter causes the neighbour entry
to be marked as permanent).

e.g.

~# evl net -S 10.10.10.10
~# arp
Address                  HWtype  HWaddress           Flags Mask            Iface
192.168.3.1              ether   3c:ec:ef:71:8e:91   C                     eth1
10.10.10.10              ether   3c:ec:ef:71:8e:91   CM                    eth1.42

This change requires ABI #38 on the kernel side.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>