latmus: add test measuring oob networking delays
This test measures the RX/TX delays when a packet traverses the oob
netstack.

'rx sched' is the propagation delay observed from the NIC driver to
the UDP entry point, this measures the latency of the RX thread in
pulling a packet before delivering it to the protocol layer
(e.g. UDP). 'rx user' is the propagation delay from the driver to the
UDP delivery point, where a packet is queued after processing to a
receiving socket, waiting for a userland consumer to dequeue
it. Therefore, 'rx user' includes 'rx sched'.

Conversely, 'tx sched' is the propagation delay observed from the
point in time when a (UDP) packet sent from userland is built until it
is queued for the TX thread to pick it later on (i.e. qdisc). 'tx
user' is the propagation delay from the UDP handling point where a
packet is received from userland until it is passed to the NIC
driver. Therefore, 'tx user' includes 'tx sched'.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>