build: reinstate some form of UAPI installation
This patch partially reverts #709e0d7 by enabling back the export of
the UAPI files to $DESTDIR/include/uapi/evl if the -Dinstall-uapi=true
is given. Typically, a user would then add -I$DESTDIR/include/uapi to
the set of CFLAGS to reach those files.

NOTE: the preferred (and saner) way to access the EVL UAPI is via
proper inclusion in the toolchain support files in the first place,
typically as part of a Yocto-generated SDK. Using a set of UAPI files
manually exported this way implicitly assumes that other UAPI files
found on the build system would be compatible with such set, which may
trigger subtle issues. You have been warned.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>