latmus: net: (strongly) report about degraded mode
In case the NIC driver is not oob-capable, make sure $user gets the
memo about the irrelevance of the latency figures plotted and/or
displayed. Extend this notice to all tests when spurious switches have
been detected.

At this chance, fix the off-by-one error with reporting the peak time
for the networking test specifically.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>