latmus: fix data collection and dump of histograms
Yet another (set of) regression(s) introduced by the recent overhaul,
with the addition of the networking test.

Several bugs mainly due to erratic code movements on this occasion
totally broke the data collection and dump of histograms. These issues
are so silly that they don't even deserve more explanation.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>