build: enable UAPI installation by default
Most users want this to happen implicitly, since the ABI files are not
provided elsewhere (e.g. by the cross-compiler toolchain, from a Yocto
SDK etc). People who don't are supposed to be aware of this tweak, so
could turn it off when needed anyway.

POLA: turn on the UAPI copy flag by default.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>