lib/net: refine error reporting on bad interface/fd
Receiving ENOENT when trying to open the /net device basically means
that CONFIG_EVL_NET is disabled in the kernel. Likewise, receiving
ENOTTY on a valid request supposedly sent to an oob port means that we
are not talking to a valid port, as a matter of fact.

Map the error codes received in these cases to ENOTSUP and EBADF
respectively.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>