tests: sched-quota-accuracy: fix yield point determination
Each quota thread needs to yield the CPU periodically in order to
ensure the correctness of the test. Fix the maths in determining the
number of loops before yielding.

Obtaining NaN as the test accuracy outcome has been observed in some
test configuration before this fix went in.

At this chance, improve some trace output as well.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>